// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/reporting/shared/dto/bill-wise-sales-report.dto.ts
export class BillWiseSalesReport_DTO {
    TransactionDate: string = '';
    BillingTransactionId: number = 0;
    PatientId: number = 0;
    PatientVisitId: number = 0;
    InvoiceNo: string = '';
    RefInvoiceNum: string = '';
    VisitType: string = '';
    ShortName: string = '';
    PatientCode: string = '';
    SchemeId: number = 0;
    SubTotal: number = 0;
    DiscountPercent: number = 0;
    DiscountAmount: number = 0;
    TotalAmount: number = 0;
    ReceivedAmount: number = 0;
    SchemeName: string = '';
    PaymentMode: string = '';
    TransactionType: string = '';
}